--- build.sh
@@ -1,4 +1,0 @@
-#!/bin/bash - 
-set -o nounset                              # Treat unset variables as an error
-cat src/*.js > media-recorder.js
-
