+++ lib/adapters.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = {
+  'flash': require('adapters/flash'),
+  'recorderjs': require('adapters/recorderjs')
+};
