+++ lib/adapters/html5.js
